% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_class.R
\name{format_class}
\alias{format_class}
\title{Transpose a classification object}
\usage{
format_class(class)
}
\arguments{
\item{class}{data.frame; a single data.frame holding the data.frame as 
returned by \code{\link[taxize]{classification}}.}
}
\value{
a wide data.frame.
}
\description{
Transpose a classification object
}
\note{
This function can only take one classification (no list of classifications), 
see examples.
}
\examples{
\dontrun{
data(samp)
# clean taxa names
require(taxize)
class <- classification(names(samp), db = 'itis')
format_class(class[[1]])
 format_class(class[[13]])
# many classications
require(plyr)
rbind.fill(lapply(class, format_class))
}
}

