% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_hier}
\alias{get_hier}
\title{Add classification information to community data}
\usage{
get_hier(x, taxa.var = "NULL", db = "itis", ...)
}
\arguments{
\item{x}{data.frame; community data. Nore this must be in transposed wide format.
(taxa in rows, samples in columns). See examples.}

\item{db}{character; database to retrieve classification data.
See  \code{\link[taxize]{classification}} for more details.}

\item{taxa.var}{character; column of taxon names.}

\item{...}{Other arguments passed to \code{\link[taxize]{classification}},}
}
\value{
A list of class wide_class, with the following elements:
\itemize{
 \item comm - community data.matrix
 \item hier - taxonomic hierarchy
 \item taxa.var - column of taxon names.
}
}
\description{
Add classification information to community data
}
\examples{
\dontrun{
data(samp)
# transpose data
df <- data.frame(t(samp), stringsAsFactors = FALSE)
df[ , 'taxon'] <- rownames(df)

# clean taxa names
get_hier(df, taxa.var = 'taxon', db = 'itis')
}
}

