% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcwp_g}
\alias{mcwp_g}
\title{Merge Children with Parents - Group variant}
\usage{
mcwp_g(x, value.var = NULL, group = NULL, level = "Family")
}
\arguments{
\item{x}{list; An object of class wide_class as returned by
\code{\link[restax]{get_hier}}.}

\item{value.var}{character; Name of the column holding the abundances to resolve}

\item{group}{character; Names of columns (=samples) for grouping}

\item{level}{character; Taxonomic level above taxa will be eliminated.
Should be returned by \code{\link[restax]{get_hier}}, see hnames therein.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item action - what was done with the taxon
 \item merged - is the taxon merged
 \item method - method to resolve taxa
}
}
\description{
Merge Children with Parents - Group variant
}
\examples{
\dontrun{
data(samp)
# transpose data
df <- data.frame(t(samp), stringsAsFactors = FALSE)
df[ , 'taxon'] <- rownames(df)
df_w <- get_hier(df, taxa.var = 'taxon', db = 'itis')
names(df_w$hier)
mcwp_g(df_w, value.var = 'S3', group = c('S1', 'S2', 'S3', 'S4'), level = 'Family')
mcwp_g(df_w, value.var = 'S3', group = c('S1', 'S2', 'S3', 'S4'), level = 'Order')
mcwp_g(df_w, value.var = 'S3', group = c('S1', 'S2', 'S3', 'S4'), level = 'Class')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

