% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restax.R
\docType{package}
\name{restax}
\alias{restax}
\alias{restax-package}
\title{A R package to resolve ambiguous taxa}
\description{
A R package to resolve ambiguous taxa, implementing the methods described in
Cuffney et al. (2007).
}
\author{
Eduard Szoecs \email{eduardszoecs@gmail.com}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M. 
Ambiguous taxa: effects on the characterization and interpretation of 
invertebrate assemblages. 
Journal of the North American Benthological Society 26, 286-307 (2007).
}

