% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rpkc_g}
\alias{rpkc_g}
\title{Remove Parents Keep Children - Group variant}
\usage{
rpkc_g(x, value.var = NULL, group = NULL, option = c("C", "K", "L"))
}
\arguments{
\item{x}{list; An object of class wide_class as returned by
\code{\link[restax]{get_hier}}.}

\item{value.var}{character; Which sampledf_w$ should be resolved.
If \code{NULL} all samples are resolved.}

\item{group}{character; Names of columns (=samples) for grouping.
If \code{NULL} all samples are grouped.}

\item{option}{character; Currently only C and L options are supported.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item action - what was done with the taxon
 \item merged - is the taxon merged
}
}
\description{
Remove Parents Keep Children - Group variant
}
\examples{
\dontrun{
data(samp)
df <- samp[1:4, ]
df <- data.frame(t(df), stringsAsFactors = FALSE)
df[ , 'taxon'] <- rownames(df)
df_w <- get_hier(df, taxa.var = 'taxon', db = 'itis')
rpkc_g(df_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), option = 'L')
# group all samples
rpkc_g(df_w, value.var = 'S2', option = 'L')
# all samples at once
rpkc_g(df_w, option = 'L')
rpkc_g(df_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), option = 'C')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

