% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rpkc_s}
\alias{rpkc_s}
\title{Remove Parents Keep Children - One sample variant.}
\usage{
rpkc_s(x, value.var = NULL)
}
\arguments{
\item{x}{list; An object of class wide_class as returned by
\code{\link[restax]{get_hier}}.}

\item{value.var}{character; Name of the column holding the abundances.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data (one sample)
 \item action - what was done with the taxon
 \item merged - is the taxon merged
 \item method - method to resolve taxa
}
}
\description{
Remove Parents Keep Children - One sample variant.
}
\examples{
\dontrun{
data(samp)
# transpose data
df <- data.frame(t(samp), stringsAsFactors = FALSE)
df[ , 'taxon'] <- rownames(df)
df_w <- get_hier(df, taxa.var = 'taxon', db = 'itis')
rpkc_s(df_w, value.var = 'A')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

