% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rpmc_g}
\alias{rpmc_g}
\title{Remove Parent or Merge Child (RPMC) - Group variant.}
\usage{
rpmc_g(x, value.var = NULL, group = NULL)
}
\arguments{
\item{x}{list; An object of class wide_class as returned by
\code{\link[restax]{get_hier}}.}

\item{value.var}{character; Which sampledf_w$ should be resolved.
If \code{NULL} all samples are resolved.}

\item{group}{character; Names of columns (=samples) for grouping.
If \code{NULL} all samples are grouped.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item action - what was done with the taxon
 \item merged - is the taxon merged
 \item method - method to resolve taxa
}
}
\description{
Remove Parent or Merge Child (RPMC) - Group variant.
}
\examples{
\dontrun{
data(samp)
# transpose data
df <- data.frame(t(samp), stringsAsFactors = FALSE)
df[ , 'taxon'] <- rownames(df)
df_w <- get_hier(df, taxa.var = 'taxon', db = 'itis')
rpmc_g(df_w, value.var = 'S3', group = c('S1', 'S2', 'S3', 'S4'))
# using all samples for grouping
rpmc_g(df_w, value.var = 'S3')
# All samples at one time
rpmc_g(df_w)
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

