% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restax.R
\name{samp}
\alias{samp}
\title{Example data}
\description{
Fictitious Example Data for restax from Cuffney et al. (2007)
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M. 
Ambiguous taxa: effects on the characterization and interpretation of 
invertebrate assemblages. 
Journal of the North American Benthological Society 26, 286-307 (2007).
}

