% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wide_class}
\alias{wide_class}
\title{Add classification information to community data}
\usage{
wide_class(comm, db = "itis", ...)
}
\arguments{
\item{comm}{data.frame; community data in wide format
(samples in rows, species in columns).}

\item{db}{character; database to retrieve classification data.
See  \code{\link[taxize]{classification}} for more details.}

\item{...}{Other arguments passed to \code{\link[taxize]{classification}},}
}
\value{
A list of class wide_class, with the following elements:
\itemize{
 \item tcomm - the transposed datamatirx with added taxonomic information
 \item hnames - the names of the columns added
}
}
\description{
Add classification information to community data
}
\examples{
\dontrun{
data(samp)
# clean taxa names
wide_class(samp)
}
}

