% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcwp_g}
\alias{mcwp_g}
\title{Merge Children with Parents - Group variant}
\usage{
mcwp_g(x, value.var = NULL, group = NULL, level = "Family")
}
\arguments{
\item{x}{list; An object of class wide_class as returned by \code{\link[restax]{wide_class}}}

\item{value.var}{character; Name of the column holding the abundances to resolve}

\item{group}{character; Names of columns (=samples) for grouping}

\item{level}{character; Taxonomic level above taxa will be eliminated.
Should be returned by \code{\link[restax]{wide_class}}, see hnames therein.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item ambp - is the taxon an ambiguous parent?
 \item ambc - is the taxon an ambiguous child?
}
}
\description{
Merge Children with Parents - Group variant
}
\examples{
\dontrun{
data(samp)
samp_w <- wide_class(samp)
samp_w$hnames
mcwp_g(samp_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), level = 'Family')
mcwp_g(samp_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), level = 'Order')
mcwp_g(samp_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), level = 'Class')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

