% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcwp_s}
\alias{mcwp_s}
\title{Merge Children with Parents - One sample variant.}
\usage{
mcwp_s(x, value.var = NULL, level = "Family")
}
\arguments{
\item{x}{list; An object of class wide_class as returned by
\code{\link[restax]{wide_class}}.}

\item{value.var}{character; Name of the column holding the abundances.}

\item{level}{character; Taxonomic level above taxa will be eliminated.
Should be returned by \code{\link[restax]{wide_class}}, see hnames therein.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item action - what was done with the taxon
 \item merged - is the taxon merged
}
}
\description{
Merge Children with Parents - One sample variant.
}
\examples{
\dontrun{
data(samp)
samp_w <- wide_class(samp)
samp_w$hnames
mcwp_s(samp_w, value.var = 'A', level = 'Family')
mcwp_s(samp_w, value.var = 'A', level = 'Order')
mcwp_s(samp_w, value.var = 'A', level = 'Class')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

