% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rpkc_g}
\alias{rpkc_g}
\title{Remove Parents Keep Children - Group variant}
\usage{
rpkc_g(x, value.var = NULL, group = NULL, option = c("C", "K", "L"))
}
\arguments{
\item{x}{list; An object of class wide_class as returned by
\code{\link[restax]{get_hier}}.}

\item{value.var}{character; Name of the column holding the abundances to resolve.}

\item{group}{character; Names of columns (=samples) for grouping.}

\item{option}{character; Currently only C and L options are supported.}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item action - what was done with the taxon
 \item merged - is the taxon merged
}
}
\description{
Remove Parents Keep Children - Group variant
}
\examples{
\dontrun{
data(samp)
samp_w <- wide_class(samp)
rpkc_g(samp_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), option = 'L')
rpkc_g(samp_w, value.var = 'S2', group = c('S1', 'S2', 'S3', 'S4'), option = 'C')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

