% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rpkc_s}
\alias{rpkc_s}
\title{Remove Parents Keep Children - One sample variant.}
\usage{
rpkc_s(x, value.var = NULL)
}
\arguments{
\item{x}{list; An object of class wide_class as returned by \code{\link[restax]{wide_class}}}

\item{value.var}{character; Name of the column holding the abundances}
}
\value{
a list of class 'restax', with the following elements
\itemize{
 \item comm - resolved community data matrix in wide format.
 \item ambp - is the taxon an ambiguous parent?
 \item ambc - is the taxon an ambiguous child?
}
}
\description{
Remove Parents Keep Children - One sample variant.
}
\examples{
\dontrun{
data(samp)
# clean taxa names
names(samp) <- gsub(' sp.', '', names(samp))
samp_w <- wide_class(samp)
rpkc_s(samp_w, value.var = 'A')
}
}
\references{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M.
Ambiguous taxa: effects on the characterization and interpretation of
invertebrate assemblages.
Journal of the North American Benthological Society 26, 286-307 (2007).
}

