% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{samp}
\alias{samp}
\alias{samp-package}
\title{Resolving ambiguous taxa}
\source{
Cuffney, T. F., Bilger, M. D. & Haigler, A. M. Ambiguous taxa: effects on the characterization and interpretation of invertebrate assemblages. Journal of the North American Benthological Society 26, 286-307 (2007).
}
\description{
Resolving ambiguous taxa

Fictitious Example Data for restax from Cuffney et al. (2007)
}
\author{
Eduard Szoecs \email{eduardszoecs@gmail.com}
}

